完成菜单、权限、角色相关功能

+++ src/app/controller/admin/permission.ts
@@ -6,9 +6,10 @@ import {
 import MyError from '@/app/common/my-error'
 import { PermissionService } from '@/app/service/permission'
 import { PermissionValidator } from '@/app/validator/permission'
+import { AdminPermissionInfo } from '@/app/model/admin-permission'
 
 @provide()
-@controller('/permission')
+@controller('/admin/permission')
 export class PermissionController {
 
   @inject('PermissionService')
@@ -32,6 +33,9 @@ export class PermissionController {
     // 校验提交的参数
     const query = this.validator.showPermission(ctx.request.query)
 
+    // 检查权限是否存在
+    await this.service.checkPermissionExists([query.id])
+
     const result = await this.service.getAdminPermissionById(query.id)
 
     ctx.helper.success(ctx, result)
@@ -44,8 +48,8 @@ export class PermissionController {
 
     const result = await this.service.createAdminPermission({
       ...params,
-      httpMethod: params.httpMethod.join(','),
-    })
+      httpMethod: params.httpMethod?.join(','),
+    } as unknown as AdminPermissionInfo)
 
     ctx.helper.success(ctx, result, null, 201)
   }
@@ -54,8 +58,14 @@ export class PermissionController {
   public async update(ctx: Context): Promise<void> {
     // 校验提交的参数
     const { id, ...params } = this.validator.updatePermission(ctx.request.body)
-    params.httpMethod = params.httpMethod?.join(',')
-    const [total] = await this.service.updateAdminPermission(id, params)
+
+    // 检查权限是否存在
+    await this.service.checkPermissionExists([id as string])
+
+    const [total] = await this.service.updateAdminPermission(id as string, {
+      ...params,
+      httpMethod: params.httpMethod?.join(','),
+    } as unknown as AdminPermissionInfo)
     assert(total, new MyError('更新失败', 400))
 
     ctx.helper.success(ctx, null, null, 204)
@@ -65,6 +75,9 @@ export class PermissionController {
   public async remove(ctx: Context): Promise<void> {
     // 校验提交的参数
     const params = this.validator.removePermission(ctx.request.body)
+
+    // 检查权限是否存在
+    await this.service.checkPermissionExists(params.ids)
 
     const total = await this.service.removeAdminPermissionByIds(params.ids)
     assert(total, new MyError('删除失败', 400))
